Moved selenium tests base-class code to mixins
